requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
fpdf2
pymupdf